Update History view to be easier to understand
